requirements.txt
matplotlib==3.6.2
pandas==1.5.2
pytest==7.2.1
python-dotenv==0.21.0
ratelimit==2.2.1
requests==2.28.1
tabulate==0.9.0
wakepy==0.5.0
web3==5.31.3
websockets==10.4
